+++ frontend/src/components/Icon.svelte
