+++ frontend/src/components/PhotoCard.svelte
@@ -5,7 +5,7 @@
   import { formatDate, formatFileSize, getThumbnailUrl, getPhotoUrl } from '../lib/utils.js';
   import { toDataURL } from '../lib/blurhash.js';
   import { t } from '../lib/i18n.js';
-  import Icon from '../lib/Icon.svelte';
+  import Icon from './Icon.svelte';
   import { longpress } from '../lib/longpress.js';
 
   const {
