+++ frontend/src/components/ViewerMetadataEdit.svelte
@@ -4,7 +4,7 @@
   import { t } from '../lib/i18n.js';
   import { addToast } from '../lib/state.svelte.js';
   import { isFormatSupported } from '../lib/utils.js';
-  import Icon from '../lib/Icon.svelte';
+  import Icon from './Icon.svelte';
 
   const { photo = null, onClose = () => {}, onSaved = () => {} } = $props();
